refactor: 그룹 정보 수정 컴포넌트 분리
