+++ src/components/domain/ManageGroup/index.jsx
@@ -1,2 +1,3 @@
 export { default as ManageMember } from './ManageMember';
 export { default as GroupDelete } from './GroupDelete';
+export { default as UpdateGroupInfo } from './UpdateGroupInfo';
